+++ xenon-common/src/main/java/com/vmware/xenon/common/AuthorizationSetupHelper.java
@@ -492,11 +492,19 @@ public class AuthorizationSetupHelper {
                         setupUser();
                         return;
                     }
-                    UserGroupState groupResponse = op.getBody(UserGroupState.class);
-                    this.userGroupSelfLink = normalizeLink(UserGroupService.FACTORY_LINK,
-                            groupResponse.documentSelfLink);
+
+                    if (op.getStatusCode() == Operation.STATUS_CODE_NOT_MODIFIED) {
+                        // user group already exists
+                        this.userGroupSelfLink = normalizeLink(UserGroupService.FACTORY_LINK, this.userGroupSelfLink);
+                    } else {
+                        UserGroupState groupResponse = op.getBody(UserGroupState.class);
+                        this.userGroupSelfLink = normalizeLink(UserGroupService.FACTORY_LINK,
+                                groupResponse.documentSelfLink);
+                    }
+
                     this.currentStep = UserCreationStep.UPDATE_USERGROUP_FOR_USER;
                     setupUser();
+
                 });
         addReplicationFactor(postGroup);
         this.host.sendRequest(postGroup);
@@ -603,9 +611,16 @@ public class AuthorizationSetupHelper {
                         setupUser();
                         return;
                     }
-                    ResourceGroupState groupResponse = op.getBody(ResourceGroupState.class);
-                    this.resourceGroupSelfLink = normalizeLink(ResourceGroupService.FACTORY_LINK,
-                            groupResponse.documentSelfLink);
+
+                    if (op.getStatusCode() == Operation.STATUS_CODE_NOT_MODIFIED) {
+                        // resource group already exists
+                        this.resourceGroupSelfLink = normalizeLink(ResourceGroupService.FACTORY_LINK, this.resourceGroupSelfLink);
+                    } else {
+                        ResourceGroupState groupResponse = op.getBody(ResourceGroupState.class);
+                        this.resourceGroupSelfLink = normalizeLink(ResourceGroupService.FACTORY_LINK,
+                                groupResponse.documentSelfLink);
+                    }
+
                     this.currentStep = UserCreationStep.MAKE_ROLE;
                     setupUser();
                 });
@@ -643,9 +658,15 @@ public class AuthorizationSetupHelper {
                         setupUser();
                         return;
                     }
-                    RoleState roleResponse = op.getBody(RoleState.class);
-                    this.roleSelfLink = normalizeLink(RoleService.FACTORY_LINK,
-                            roleResponse.documentSelfLink);
+
+                    if (op.getStatusCode() == Operation.STATUS_CODE_NOT_MODIFIED) {
+                        // role already exists
+                        this.roleSelfLink = normalizeLink(RoleService.FACTORY_LINK, this.roleSelfLink);
+                    } else {
+                        RoleState roleResponse = op.getBody(RoleState.class);
+                        this.roleSelfLink = normalizeLink(RoleService.FACTORY_LINK, roleResponse.documentSelfLink);
+                    }
+
                     this.currentStep = UserCreationStep.SUCCESS;
                     setupUser();
                 });