+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -3747,9 +3747,7 @@ public class ServiceHost implements ServiceRequestSender {
             }
 
             op.setStatusCode(fo.getStatusCode());
-            if (fo.hasBody()) {
-                op.setBodyNoCloning(fo.getBodyRaw());
-            }
+            op.setBodyNoCloning(fo.getBodyRaw());
 
             op.setContentType(fo.getContentType());
             op.setContentLength(fo.getContentLength());