+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestResourceGroupService.java
@@ -121,13 +121,10 @@ public class TestResourceGroupService extends BasicReusableHostTestCase {
         assertEquals(state.query.term.propertyName, responseState.query.term.propertyName);
         assertEquals(state.query.term.matchValue, responseState.query.term.matchValue);
 
-        responseState = (ResourceGroupState) this.host.verifyPost(ResourceGroupState.class,
+        this.host.verifyPost(ResourceGroupState.class,
                 ServiceUriPaths.CORE_AUTHZ_RESOURCE_GROUPS,
                 state,
                 Operation.STATUS_CODE_NOT_MODIFIED);
-
-        assertEquals(state.query.term.propertyName, responseState.query.term.propertyName);
-        assertEquals(state.query.term.matchValue, responseState.query.term.matchValue);
 
         state.query.setTermMatchValue("valueModified");